Fixed possible spurious lines for DSO outlines and constellations (precission error; fix #160)

+++ src/core/modules/Constellation.hpp
@@ -179,8 +179,8 @@ private:
 	LinearFader artFader, lineFader, nameFader, boundaryFader;
 	//! Constellation art opacity
 	float artOpacity;
-	std::vector<std::vector<Vec3f> *> isolatedBoundarySegments;
-	std::vector<std::vector<Vec3f> *> sharedBoundarySegments;
+	std::vector<std::vector<Vec3d> *> isolatedBoundarySegments;
+	std::vector<std::vector<Vec3d> *> sharedBoundarySegments;
 
 	//! Currently we only need one color for all constellations, this may change at some point
 	static Vec3f lineColor;
